move more things to funsor.typing
